+++ CHANGES.md
@@ -9,6 +9,9 @@ To be released.
  -  (Libplanet.RocksDBStore) Fixed a bug that  `RocksDBStore.DeleteChainId()`
     method had thrown `KeyNotFoundException` when there's no such chain ID.
     [[#891]]
+ -  (Libplanet.RocksDBStore) Fixed a bug that `RocksDBStore` had written logs
+    into the incorrect context `DefaultContext`, not `RocksDBStore`
+    the correct one.  [[#891]]
 
 [#891]: https://github.com/planetarium/libplanet/pull/891